# Conflicts:
#	src/Main.java
#	src/Unit.java

+++ src/OngoingBattle.java
@@ -1,8 +1,8 @@
 public class OngoingBattle {
 
     public final boolean seaBattle;
-    public int aInf, aTank, aFig, aBom, aSub, aDes, aACC, aBat, aExtra;
-    public int dInf, dTank, dFig, dBom, dSub, dDes, dACC, dBat, dExtra;
+    public byte aInf, aTank, aFig, aBom, aSub, aDes, aACC, aBat, aExtra;
+    public byte dInf, dTank, dFig, dBom, dSub, dDes, dACC, dBat, dExtra;
 
     OngoingBattle(Battle battle) {
         this.seaBattle = battle.seaBattle;
